Add calibration functionality into LunchBox device.
A few new analysis functions have been added to allow the
LunchBox ophyd device to calculate the distance between the 1st
order and middle diffraction spots found. This uses a piece of
paper with two black blobs some distance apart (assumed to be
10cm). Therefore this code will fail an assertion check if no
such blobs exist in the camera view.